refactor(construction): split constructionRules into focused modules
