+++ src/utils/construction/constructionRules/index.ts
@@ -1,0 +1,28 @@
+/**
+ * Construction Rules Barrel Export
+ */
+
+export type {
+  ConstructionResult,
+  ConstructionStepResult,
+  MechBuildConfig,
+} from './types';
+
+export {
+  calculateArmor,
+  calculateRemainingTonnage,
+  calculateStructuralWeight,
+  getEquipmentCritEntries,
+  isSuperHeavy,
+  validateTonnage,
+} from './tonnage';
+
+export {
+  calculateCockpit,
+  calculateEngine,
+  calculateGyro,
+  calculateHeatSinks,
+  calculateInternalStructure,
+} from './components';
+
+export { validateConstruction } from './validation';
